requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.9.0
boto3==1.34.155
botocore==1.34.155
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
compel==2.0.3
contourpy==1.3.0
controlnet-aux==0.0.3
cycler==0.12.1
Cython==3.0.11
dacite==1.6.0
dagshub==0.4.2
dagshub-annotation-converter==0.1.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.30.0
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.0.1
fairscale==0.4.4
fastapi==0.115.6
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.13.1
Flask==2.2.3
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
flaskwebgui==0.3.5
flatbuffers==24.12.23
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.3.1
fusepy==3.0.1
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
gradio==5.9.1
gradio_client==1.5.2
graphql-core==3.2.3
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@a52c6b59ecbe5e57d6469106a40ae0f70257a051#egg=groundingdino&subdirectory=GroundingDINO
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
imageio==2.35.1
importlib_metadata==8.2.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning-utilities==0.11.6
litellm==1.56.10
loguru==0.7.3
lxml==5.3.0
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.20.1
open-clip-torch==2.17.1
openai==1.59.3
opencv-python==4.10.0.84
orjson==3.10.13
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.0.0
pexpect==4.9.0
piexif==1.1.3
pillow==10.2.0
platformdirs==4.2.2
portalocker==3.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==2.0.7
python-multipart==0.0.20
python-socketio==5.12.1
pytorch-fid==0.3.0
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.1.0
rpds-py==0.20.0
ruff==0.8.6
s3transfer==0.10.2
safehttpx==0.1.6
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@a52c6b59ecbe5e57d6469106a40ae0f70257a051#egg=segment_anything&subdirectory=segment_anything
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sng_parser==1.3.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.41.3
supervision==0.25.0
sympy==1.13.1
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
tifffile==2024.9.20
tiktoken==0.8.0
timm==0.4.12
tinycss2==1.3.0
tokenizers==0.20.3
tomli==2.0.1
tomlkit==0.13.2
torch==2.5.1
torch-fidelity==0.3.0
torchaudio==2.5.1
torchmetrics==1.4.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
treelib==1.6.4
triton==3.1.0
typer==0.15.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.34.0
wandb==0.17.5
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==2.2.2
whichcraft==0.6.1
widgetsnbextension==4.0.11
wsproto==1.2.0
xformers==0.0.29.post1
yacs==0.1.8
yapf==0.43.0
yarl==1.9.4
zipp==3.19.2
